feat: dashboard admin - cards com estatisticas em tempo real
- DashboardComponent: 6 cards (convidados, confirmados, presentes, presenteados, valor total, fotos)
- grid responsivo 1/2/3 colunas, cards com hover suave e gradiente dourado
- tipografia via mixins _typography.scss (h1, body-lg, body-sm, h3)
- chamada GET /api/admin/stats com JWT via interceptor
- fix: GiftPurchaseRepository g.price -> g.value (BigDecimal)
- fix: environment.ts loginEncryptionKey batendo com .env
- AGENTS.md: frontend env vars documentadas

+++ backend/src/main/java/com/pedrogio/wedding/stats/StatsService.java
@@ -6,6 +6,8 @@ import com.pedrogio.wedding.gift.GiftRepository;
 import com.pedrogio.wedding.guest.GuestRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 public class StatsService {
 
@@ -29,8 +31,8 @@ public class StatsService {
         long confirmedGuests = guestRepository.countByConfirmed(true);
         long totalGifts = giftRepository.count();
         long giftedGifts = purchaseRepository.countGiftsWithPaidPurchases();
-        Long sum = purchaseRepository.sumPaidAmounts();
-        long totalGiftedValue = sum != null ? sum : 0L;
+        BigDecimal sum = purchaseRepository.sumPaidAmounts();
+        long totalGiftedValue = sum != null ? sum.longValue() : 0L;
         long totalPhotos = photoRepository.count();
 
         return new StatsResponse(
